moved stage-6 rasterization from engine context to gosub_render_pipeline
