+++ crates/gosub_engine/Cargo.toml
@@ -50,7 +50,6 @@ winit = { workspace = true, optional = true }
 
 eframe = { workspace = true, features = ["wgpu"], optional = true }
 egui = { workspace = true, optional = true }
-rayon = "1"
 time = "0.3.41"
 tokio-util = { workspace = true, features = ["io"] }
 tempfile = { workspace = true }
